replaced current path finding with cuongs

+++ include/util.h
@@ -32,7 +32,7 @@ namespace util
     @brief normalizes a vector (sqrt of dot product with itself)
     @param vector coords
   */
-  static float norm(float x, float y, float z){
+  static double norm(double x, double y, double z){
     return sqrt(x*x + y*y + z*z);
   }
 }
